requirements.txt
db-sqlite3
Pillow
streamlit